Refactor sortclaims functions into own file
